+++ sdk/typescript/test/e2e/encrypt-user-share.test.ts
@@ -17,10 +17,10 @@ import {
 import { Config } from '../../src/dwallet-mpc/globals';
 import { signWithEncryptedDWallet } from '../../src/dwallet-mpc/sign';
 import { Ed25519Keypair } from '../../src/keypairs/ed25519';
-import { mockCreateDwallet } from './utils/dwallet';
 import {
 	DKGCentralizedPrivateOutput,
 	DKGDecentralizedOutput,
+	mockCreateDwallet,
 	mockedProtocolPublicParameters,
 } from './utils/dwallet_mocks';
 import { setup, TestToolbox } from './utils/setup';
