+++ sdk/typescript/test/e2e/utils/dwallet_mocks.ts
@@ -1,5 +1,17 @@
 import { Buffer } from 'buffer';
 
+import { bcs } from '../../../src/bcs/index.js';
+import type { Config, CreatedDwallet, DWallet } from '../../../src/dwallet-mpc/globals.js';
+import {
+	dWallet2PCMPCECDSAK1ModuleName,
+	dWalletMoveType,
+	isDWallet,
+	packageId,
+} from '../../../src/dwallet-mpc/globals.js';
+import type { Presign } from '../../../src/dwallet-mpc/presign.js';
+import { isPresign, presignMoveType } from '../../../src/dwallet-mpc/presign.js';
+import { Transaction } from '../../../src/transactions/index.js';
+
 const MockedProtocolPublicParameters =
 	'OlRoZSBmaW5pdGUgZmllbGQgb2YgaW50ZWdlcnMgbW9kdWxvIHByaW1lIHEgJFxtYXRoYmJ7Wn1fcSRBQTbQjF7SvzugSK/m3K66/v///////////////////wAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABAAYAAAAAAADqAgAAAAAAAGsEAAAAAAAAi8KxSPEVhbHczB2vJ3IG0r1gwz6FRQbt34obZKohtOxf5aqrgc+Mcb1ySZQiht4Z/DMw9/2KFp0cRd8AZZZG+FhI/EDWxnA8BeINcUd8sqPkhZaHiI06ZyvD2LFAGceI3+9Y6lAR93eXwwTVJ9WLQGrmzcImQPnIshR9YuAZK2kBV4z49vNgTWMznWeEbFg6F3JV8Uj+gy4MBXyTyvVinx7ONncCaTKsy4mOD944+9C9R4/r05BzGE1lKGVQFgBPJI4IS2SSXe9eV0psbBzfmqQkU5wpj+QcrlX1FrIhsHttyRpfPHAm5wkAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAIAAAAAAAAAAQt3/QyveX85aTo4sVSI/+7Wx9ox9R8vo232sPdAS6tQ2jfmlLB7ggESmYdbYH5lAznThVtbV8iVbELATGQjn83s4sBDRbXgSWBJpuNscx5XcRwANMqiteToE6ugMV/6oQHJNnTr94YL/QVP270T2yQkPOKSQRcOjzfGHO3rC7MPqG3AR7BXD3cxITWGJTmIl6us29KfJRNIx1X1FrIhsHttyRpfPHAm5wkAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAGLwrFI8RWFsdzMHa8ncgbSvWDDPoVFBu3fihtkqiG07F/lqquBz4xxvXJJlCKG3hn8MzD3/YoWnRxF3wBllkb4WEj8QNbGcDwF4g1xR3yyo+SFloeIjTpnK8PYsUAZx4jf71jqUBH3d5fDBNUn1YtAaubNwiZA+ciyFH1i4BkraQFXjPj282BNYzOdZ4RsWDoXclXxSP6DLgwFfJPK9WKfHs42dwJpMqzLiY4P3jj70L1Hj+vTkHMYTWUoZVAWAE8kjghLZJJd715XSmxsHN+apCRTnCmP5ByuVfUWsiGwe23JGl88cCbnCQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAQUE20Ixe0r87oEiv5tyuuv7///////////////////9LGXmpYJMPBz9awXCBPs1lRxrUFqRFY8gj7WbHwh+O2rYx8aLCdyhivU85ixJoos0sjVUuG9wCHS544s4tVdRY+kDjL5gEolwpFepSPNygbnf594AKn2zZt8yKRjpNefvP01Ht8mvBtzKGKhOCx9wVr/Af7tpwqt3TVfUWsiGwe23JGl88cCbnCQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABQUE20Ixe0r87oEiv5tyuuv7///////////////////8AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAIESDjihUD9JtJjFvahzwy7FB81b5PWX5sWbxoHVHGedg4JsoBm9pH93QJFezblddf3///////////////////8AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABwAAAAAAAAAAEQgRIOOKFQP0m0mMW9qHPDLsUHzVvk9ZfmxZvGgdUcZ52DgmygGb2kf3dAkV7NuV11/f///////////////////wEIQUE20Ixe0r87oEiv5tyuuv7///////////////////8BK8NdP8Nr3l9Omg5OLBXiv7v1sXZM/cfLaJs97D3Q0iqU9o05JeyeYIBE5uEWmF/ZQM504dbWFXIlmxAwE9nIZzO7OPBQUS14EliS6TjbnMdVHAcAjbJoLTn6xCpozJd+aECyDd36veFCf8HTdu/ENgkJj7hk0IXD440xR/v6wuwDahvwEezVw11MSI1hSQ7i5SrrtvRnyQTScVW9hWwI7F5bssYXD5zJeQIAAAD/O4vCsUjxFYWx3MwdrydyBtK9YMM+hUUG7d+KG2SqIbTsX+Wqq4HPjHG9ckmUIobeGfwzMPf9ihadHEXfAGWWRvhYSPxA1sZwPAXiDXFHfLKj5IWWh4iNOmcrw9ixQBnHiN/vWOpQEfd3l8ME1SfVi0Bq5s3CJkD5yLIUfWLgGStpAVeM+PbzYE1jM51nhGxYOhdyVfFI/oMuDAV8k8r1Yp8ezjZ3AmkyrMuJjg/eOPvQvUeP69OQcxhNZShlUBYATySOCEtkkl3vXldKbGwc35qkJFOcKY/kHK5V9RayIbB7bckaXzxwJucJAAAAAR49xCff7KK0Yc6JINz/12f0sy1kxUXjzthbbPWDlvQOWE1LmTIKmCfbGXBkjo8zm5JRQSJnOVCCajWGh8J19m6xIRwjRQvyrcVhQ0T8j+PAetck0KHO+zDmtitjfVDo6JA/35eUX7xaKcjtc/qBE5F7lx4NFgEAAAABHXGWc7TIEBFB3COpZ30KGTGd4WKOf/n20xrgp4ovKPRr0pPyslCAdtt3Yu+YOysqObB8LAKrtNd97VrCxOFK9KHKJ/hZ5FT999irnQw7rMZkubZy9jgCQpqVbjXGVbivI1jaakcqd8wDozxi7KkSNTIYIME1AR43mzxd7/WQIRKPjb/GMDNgLH4CJZFu1QvLm3M7iYJXQy4dMjwm4kjy3KOp5N/PrbS+ESXL+HsGGoIbgyK21u3LPZQQqEWl3+acgP4E4fyf2NhJZP1xOZxH/P4oVTp0ICLMoRUzbJyCWHi4DJQQpDcdxEQfSgIAAAD/O4vCsUjxFYWx3MwdrydyBtK9YMM+hUUG7d+KG2SqIbTsX+Wqq4HPjHG9ckmUIobeGfwzMPf9ihadHEXfAGWWRvhYSPxA1sZwPAXiDXFHfLKj5IWWh4iNOmcrw9ixQBnHiN/vWOpQEfd3l8ME1SfVi0Bq5s3CJkD5yLIUfWLgGStpAVeM+PbzYE1jM51nhGxYOhdyVfFI/oMuDAV8k8r1Yp8ezjZ3AmkyrMuJjg/eOPvQvUeP69OQcxhNZShlUBYATySOCEtkkl3vXldKbGwc35qkJFOcKY/kHK5V9RayIbB7bckaXzxwJucJAAAAAAEepQG9PaCt/VPhzevyDO+i3z1Kh10tqAU0RYAqF57uaXBlvsy4c0I67ayuMFVpb4XE7cmgJ+QTKmGjvoNuxVGiyWaslmXAD+SOgEo+5CpyqWLao6GTvYAIkIdtFFREakuB9aaZ7Ys0yB3pCArp3ioseY4rC3wBAAAAAR3lSxrVDxHGVFazrT4mRAhvsH5YixWBAkDyBB+/JabAqfrsOYIX3p5yZ0Asw8YUYCGDs/m59uuuAI7JZ8xHZgqYodpPNiEyZo+SEd9Cjv2bVsm5pB8BAFDl2mIfUbOZIxCF8GrxWA2ox9Sc6MQELZ6E0hwWUgEepeAPyqG6zORqH28l6W2+8ADBFlbDgKCPkCXIrA8B35pKBScPjckRZJV5S1J+qRy/80GLtoTOZNwRheWAohEwtTOsCOlgb/85Xo974qJFfQbwjS5LJ2jpqdrMNq3IuLAWIQxVa+8FoDBvQfPGv8PlBK8rW68BAAAA/zuLwrFI8RWFsdzMHa8ncgbSvWDDPoVFBu3fihtkqiG07F/lqquBz4xxvXJJlCKG3hn8MzD3/YoWnRxF3wBllkb4WEj8QNbGcDwF4g1xR3yyo+SFloeIjTpnK8PYsUAZx4jf71jqUBH3d5fDBNUn1YtAaubNwiZA+ciyFH1i4BkraQFXjPj282BNYzOdZ4RsWDoXclXxSP6DLgwFfJPK9WKfHs42dwJpMqzLiY4P3jj70L1Hj+vTkHMYTWUoZVAWAE8kjghLZJJd715XSmxsHN+apCRTnCmP5ByuVfUWsiGwe23JGl88cCbnCQAAAA==';
 export const mockedProtocolPublicParameters = Uint8Array.from(
@@ -21,3 +33,105 @@ export const mockedPresign = {
 	presign: Uint8Array.from(Buffer.from(MockedPresign, 'base64')),
 	firstRoundSessionID: '0x94cb1bae7e282ad800f78e176f08ae12ddc3784c2810232c7a20ea8ab41244b8',
 };
+
+export async function mockCreateDwallet(c: Config): Promise<CreatedDwallet> {
+	console.log('Creating dWallet Mock');
+
+	// Initiate the transaction
+	const tx = new Transaction();
+	tx.moveCall({
+		target: `${packageId}::${dWallet2PCMPCECDSAK1ModuleName}::create_mock_dwallet`,
+		arguments: [
+			tx.pure(bcs.vector(bcs.u8()).serialize(mockedDWallet.decentralizedDKGOutput)),
+			tx.pure(bcs.vector(bcs.u8()).serialize(mockedDWallet.centralizedDKGPublicOutput)),
+		],
+	});
+
+	// Execute the transaction
+	const res = await c.client.signAndExecuteTransaction({
+		signer: c.keypair,
+		transaction: tx,
+		options: {
+			showEffects: true,
+		},
+	});
+
+	// Validate the created objects
+	const createdObjects = res.effects?.created;
+	if (!createdObjects || createdObjects.length !== 2) {
+		throw new Error(
+			`mockCreateDwallet error: Unexpected number of objects created. Expected 2, got ${
+				createdObjects?.length || 0
+			}`,
+		);
+	}
+	await new Promise((resolve) => setTimeout(resolve, 2000));
+	for (const obj of createdObjects) {
+		const objectData = await c.client.getObject({
+			id: obj.reference.objectId,
+			options: { showContent: true },
+		});
+		const dwalletData =
+			objectData.data?.content?.dataType === 'moveObject' &&
+			objectData.data?.content.type === dWalletMoveType &&
+			isDWallet(objectData.data.content.fields)
+				? (objectData.data.content.fields as DWallet)
+				: null;
+
+		if (dwalletData) {
+			return {
+				id: dwalletData.id.id,
+				centralizedDKGPublicOutput: Array.from(Buffer.from(DKGCentralizedPublicOutput, 'base64')),
+				centralizedDKGPrivateOutput: Array.from(Buffer.from(DKGCentralizedPrivateOutput, 'base64')),
+				decentralizedDKGOutput: dwalletData.decentralized_output,
+				dwalletCapID: dwalletData.dwallet_cap_id,
+				dwalletMPCNetworkKeyVersion: dwalletData.dwallet_mpc_network_key_version,
+			};
+		}
+	}
+	throw new Error(`mockCreateDwallet error: failed to create an object of type ${dWalletMoveType}`);
+}
+
+export async function mockCreatePresign(c: Config, dwallet: CreatedDwallet): Promise<Presign> {
+	console.log('Creating Presign Mock');
+	const tx = new Transaction();
+	const [presign] = tx.moveCall({
+		target: `${packageId}::${dWallet2PCMPCECDSAK1ModuleName}::create_mock_presign`,
+		arguments: [
+			tx.pure.id(dwallet.id),
+			tx.pure(bcs.vector(bcs.u8()).serialize(mockedPresign.presign)),
+			tx.pure.id(mockedPresign.firstRoundSessionID),
+		],
+	});
+	tx.transferObjects([presign], c.keypair.toPeraAddress());
+	let res = await c.client.signAndExecuteTransaction({
+		signer: c.keypair,
+		transaction: tx,
+		options: {
+			showEffects: true,
+		},
+	});
+	const presignID = res.effects?.created?.at(0)?.reference.objectId;
+	if (!presignID) {
+		throw new Error('create_mock_presign error: Failed to create presign');
+	}
+	await new Promise((resolve) => setTimeout(resolve, 2000));
+	const obj = await c.client.getObject({
+		id: presignID,
+		options: { showContent: true },
+	});
+	const preSignObj =
+		obj.data?.content?.dataType === 'moveObject' &&
+		obj.data?.content.type === presignMoveType &&
+		isPresign(obj.data.content.fields)
+			? (obj.data.content.fields as Presign)
+			: null;
+
+	if (!preSignObj) {
+		throw new Error(
+			`invalid object of type ${dWalletMoveType}, got: ${JSON.stringify(obj.data?.content)}`,
+		);
+	}
+
+	return preSignObj;
+}
